Move util into shared.
