+++ lsif/src/server/backend/backend.ts
@@ -1,6 +1,6 @@
 import * as lsp from 'vscode-languageserver-protocol'
 import { ConnectionCache, DocumentCache, ResultChunkCache } from './cache'
-import { dbFilename, mustGet, NoLSIFDumpError } from '../../util'
+import { dbFilename, mustGet, NoLSIFDumpError } from '../../shared/util'
 import { XrepoDatabase } from '../../shared/xrepo/xrepo'
 import { TracingContext, logAndTraceCall, addTags, logSpan } from '../../shared/tracing'
 import { Database, sortMonikers, createRemoteUri } from './database'
