+++ lsif/src/shared/util.ts
@@ -1,8 +1,8 @@
 import * as fs from 'mz/fs'
 import * as path from 'path'
 import { Id } from 'lsif-protocol'
-import * as constants from './constants'
-import { DefinitionReferenceResultId } from './shared/models/types'
+import * as constants from '../constants'
+import { DefinitionReferenceResultId } from './models/types'
 
 /**
  * Reads an integer from an environment variable or defaults to the given value.
