+++ lsif/src/worker/processors/convert.ts
@@ -4,7 +4,7 @@ import * as path from 'path'
 import { ConfigurationFetcher } from '../../shared/config/config'
 import { convertLsif } from '../importer/importer'
 import { createSilentLogger } from '../../shared/logging'
-import { dbFilename, hasErrorCode } from '../../util'
+import { dbFilename, hasErrorCode } from '../../shared/util'
 import { DBS_DIR_MAXIMUM_SIZE_BYTES, STORAGE_ROOT } from '../settings'
 import { logAndTraceCall, TracingContext } from '../../shared/tracing'
 import { XrepoDatabase } from '../../shared/xrepo/xrepo'
