+++ lsif/src/worker/worker.ts
@@ -9,7 +9,7 @@ import { createLogger } from '../shared/logging'
 import { createPostgresConnection } from '../shared/database/postgres'
 import { createQueue } from '../shared/queue/queue'
 import { createUpdateTipsJobProcessor } from './processors/update-tips'
-import { ensureDirectory } from '../util'
+import { ensureDirectory } from '../shared/util'
 import { followsFrom, FORMAT_TEXT_MAP, Span, Tracer } from 'opentracing'
 import { instrument } from '../shared/metrics'
 import { Job } from 'bull'
